chore: move `HTTPClientMock` to its own file
