+++ SmartCityExploration/SmartCityExplorationTests/API/CitiesAPIClientTests.swift
@@ -132,15 +132,3 @@ final class CitiesAPIClientTests: XCTestCase {
     }
 
 }
-
-private final class HTTPClientMock: HTTPClient {
-    var clientResponse: HTTPClient.Result!
-
-    func get(from url: URL) async -> HTTPClient.Result {
-        guard let clientResponse else {
-            fatalError("Developer error, mock response not set.")
-        }
-
-        return clientResponse
-    }
-}
